+++ vizro-core/pyproject.toml
@@ -57,7 +57,7 @@ exclude_lines = [
   "if __name__ == .__main__.:",
   "if TYPE_CHECKING:"
 ]
-fail_under = 90
+fail_under = 91
 show_missing = true
 skip_covered = true
